Adding functionnality of sharing Media through specifics applications
