--- userprofile/src/main/java/com/tripbook/userprofileManfoIngrid/domain/model/MediaItem.kt
@@ -1,25 +1,0 @@
-package com.tripbook.userprofileManfoIngrid.domain.model
-
-import java.util.Date
-
-data class MediaItem(
-    val id: String,
-    val name: String,
-    val type: MediaType,
-    val url: String,
-    val createdDate: Date,
-    val modifiedDate: Date,
-    val size: Long
-)
-
-enum class MediaType {
-    PHOTO, VIDEO
-}
-
-enum class MediaFilter {
-    ALL, PHOTOS, VIDEOS
-}
-
-enum class SortOption {
-    CREATION_DATE, MODIFICATION_DATE, NAME
-}
