+++ userprofile/src/main/java/com/tripbook/userprofileManfoIngrid/presentation/MediaOrganizationScreen.kt
@@ -6,7 +6,6 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import com.tripbook.userprofileManfoIngrid.presentation.media.components.MediaGrid
