+++ userprofile/src/main/java/com/tripbook/userprofileManfoIngrid/presentation/media/components/dialog/EditMediaDialog.kt
@@ -1,2 +1,65 @@
-package com.tripbook.userprofileManfoIngrid.presentation.media.components.dialog
+package com.tripbook.userprofileManfoIngrid.presentation.media.components.dialogs
 
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Dialog
+import com.tripbook.userprofileManfoIngrid.presentation.media.models.MediaItem
+
+@Composable
+fun EditMediaDialog(
+    media: MediaItem,
+    onDismiss: () -> Unit,
+    onSave: (String) -> Unit
+) {
+    var newName by remember { mutableStateOf(media.name) }
+
+    Dialog(onDismissRequest = onDismiss) {
+        Card(
+            shape = RoundedCornerShape(16.dp),
+            modifier = Modifier.padding(16.dp)
+        ) {
+            Column(
+                modifier = Modifier.padding(20.dp)
+            ) {
+                Text(
+                    text = "Modifier le nom",
+                    fontSize = 18.sp,
+                    fontWeight = FontWeight.Bold
+                )
+
+                Spacer(modifier = Modifier.height(16.dp))
+
+                OutlinedTextField(
+                    value = newName,
+                    onValueChange = { newName = it },
+                    label = { Text("Nouveau nom") },
+                    modifier = Modifier.fillMaxWidth()
+                )
+
+                Spacer(modifier = Modifier.height(20.dp))
+
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.SpaceEvenly
+                ) {
+                    TextButton(onClick = onDismiss) {
+                        Text("Annuler")
+                    }
+
+                    Button(
+                        onClick = { onSave(newName) },
+                        enabled = newName.isNotBlank()
+                    ) {
+                        Text("OK")
+                    }
+                }
+            }
+        }
+    }
+}
