+++ userprofile/src/main/java/com/tripbook/userprofileManfoIngrid/presentation/media/models/MediaItem.kt
@@ -1,4 +1,25 @@
 package com.tripbook.userprofileManfoIngrid.presentation.media.models
 
-class MediaItem {
+import java.util.Date
+
+data class MediaItem(
+    val id: String,
+    val name: String,
+    val type: MediaType,
+    val url: String,
+    val createdDate: Date,
+    val modifiedDate: Date,
+    val size: Long
+)
+
+enum class MediaType {
+    PHOTO, VIDEO
+}
+
+enum class MediaFilter {
+    ALL, PHOTOS, VIDEOS
+}
+
+enum class SortOption {
+    CREATION_DATE, MODIFICATION_DATE, NAME
 }
